Significantly more dotfiles.
An oceanic tide of existing dotfiles are now dutifully archived, the
most noteworthy of which is the antiquated "perl/cp-pcre" Perl script:
a GPL-licensed variant on the "cp" and "mv" coreutils permitting
arbitrarily many pathnames and/or plaintext file contents to be
recursively munged via a Perl-compatible Regular Expression (PCRE),
complete with full-blown substitutions and dry run support.

This script is sufficiently robust to warrant inclusion in CPAN. Until
then, it resides here.